Sequence expr evaluation
Promise.all rejects at first child rejection, this unfortunately leaves
the runtime in an incomplete state when returning if a leaf is
rejected, as other expression evaluations may still be in-flight
at the time of rejection. There is a way around this, but it'll
require more substantial refactoring of the evaluator as well as
the expr semantics.

This unfortunately makes provisioning slower, but at least we'll have
reliable runtime results upon failure.